requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
pandas>=2.0.0
